llm-service/requirements.txt
fastapi
uvicorn
python-dotenv
requests
# Requerimientos para el modelo LLM
transformers>=4.51.0
torch
accelerate
safetensors
sentencepiece